+++ contrib/scripts/generate-export-realm.sh
@@ -14,14 +14,15 @@ touch "${GIT_ROOT}/src/test/resources/import-files/exported-realm/${KEYCLOAK_VER
 
 docker run -d --rm \
   --name keycloak-export \
-  -v "${MKTMP}:/opt/jboss/keycloak/standalone/data" \
+  -v "${MKTMP}:/opt/keycloak/imex" \
   -e ROOT_LOGLEVEL=ERROR \
   -e KEYCLOAK_LOGLEVEL=ERROR \
-  -e KEYCLOAK_USER=admin \
-  -e KEYCLOAK_PASSWORD=admin123 \
-  "quay.io/keycloak/keycloak:${KEYCLOAK_VERSION}-legacy"
+  -e KEYCLOAK_ADMIN=admin \
+  -e KEYCLOAK_ADMIN_PASSWORD=admin123 \
+  "quay.io/keycloak/keycloak:${KEYCLOAK_VERSION}" \
+  start-dev
 
-while ! docker exec keycloak-export bash -c '/opt/jboss/keycloak/bin/kcadm.sh config credentials --server http://$HOSTNAME:8080/auth --realm master --user $KEYCLOAK_USER --password $KEYCLOAK_PASSWORD'; do
+while ! docker exec keycloak-export bash -c '/opt/keycloak/bin/kcadm.sh config credentials --server http://$HOSTNAME:8080/auth --realm master --user $KEYCLOAK_USER --password $KEYCLOAK_PASSWORD'; do
   sleep 2
 done
 
@@ -34,13 +35,13 @@ sleep 2
 docker run -ti --rm \
   --name keycloak-export \
   -v "${GIT_ROOT}/src/test/resources/import-files/exported-realm/${KEYCLOAK_VERSION}/:/tmp/export/" \
-  -v "${MKTMP}:/opt/jboss/keycloak/standalone/data" \
+  -v "${MKTMP}:/opt/keycloak/imex" \
   -e ROOT_LOGLEVEL=ERROR \
   -e KEYCLOAK_LOGLEVEL=ERROR \
   -e KEYCLOAK_USER=admin \
   -e KEYCLOAK_PASSWORD=admin123 \
   --entrypoint bash \
-  "quay.io/keycloak/keycloak:${KEYCLOAK_VERSION}-legacy" \
-  -c "timeout 30 /opt/jboss/keycloak/bin/standalone.sh -Dkeycloak.migration.action=export -Dkeycloak.migration.provider=dir -Dkeycloak.migration.dir=/tmp/export -Dkeycloak.migration.usersExportStrategy=REALM_FILE"
+  "quay.io/keycloak/keycloak:${KEYCLOAK_VERSION}" \
+  -c "timeout 30 /opt/keycloak/bin/kc.sh export --dir /tmp/export --users realm_file"
 
 rm -rf "${MKTMP}"